requirements.txt
python-decouple
"psycopg[binary]"